feat(game_task): port CGameTask::HandleMessage + name base fields
Pure-Rust port of CGameTask::HandleMessage (0x004FF280, vtable slot 2)
with all 9 WA call sites hooked through a thiscall trampoline. The
explicit message cases in WA's original are mirrored verbatim:

- 0x02 (FrameFinish): opportunistic GC of an owned sound handle —
  poll the sound subsystem via the renamed CGameTask::IsSoundHandleExpired
  helper (return value semantics were the inverse of the original
  Ghidra label) and call CGameTask::ReleaseSoundHandle when the slot
  has been reclaimed.
- 0x1c (Explosion): clamp health to >=0, run CGameTask::ComputeExplosionDamage
  to get the per-entity falloff-adjusted damage, accumulate into the
  new damage_accum field, and on caller_flag report a percentage to
  CTaskTurnGame via msg 0x1d.
- 0x4b: clamp health, then dispatch self vtable slot 17 with two
  payload words.
- default: CTask::broadcast_message_raw.

Three Ghidra functions renamed for clarity: FUN_004FF390 ->
CGameTask::ComputeExplosionDamage (usercall, EDI=this), FUN_00546CD0 ->
CGameTask::IsSoundHandleExpired (usercall, ECX=this + EAX=handle),
FUN_00546D20 -> CGameTask::ReleaseSoundHandle (fastcall ECX/EDX).
Initial bridge for IsSoundHandleExpired was thiscall-only, missing the
EAX parameter; the fix is latent (no test scenarios exercise an active
this->sound_handle) but headful audio testing confirmed no regressions.

Two CGameTask base fields split out of _unknown_98:
- damage_accum (i32, +0xD4): per-entity damage accumulator, written
  by msg=0x1c.
- sound_handle (u32, +0xE0): owned sound emitter handle, polled and
  cleared by msg=0x02.

The "health" field at +0x48 stays a local OFFSET_HEALTH constant since
it lives in subclass_data. Subclasses that delegate to CGameTask::HandleMessage
(mines, oil drums, gravestone, score bubbles) treat it as HP; subclasses
that override HandleMessage outright (notably CTaskWorm, which uses the
slot as set_action_field_raw) never reach this path.

CTaskTurnGame::from_shared_data + SHARED_DATA_KEY added so the
per-game root task can be located without naming the (0, 0x14) magic
key at every call site. create_explosion drops the manual SharedDataTable
+ vtable transmute in favor of the typed dispatch.

Includes a partial WIP from a separate refactor toward typed
TaskMessage args; the tree does not currently build standalone but
all explosion work passes 15/15 headless on the prior buildable commit.

+++ crates/openwa-game/src/game/create_explosion.rs
@@ -67,7 +67,7 @@ unsafe fn dispatch(sender: *mut CTask, msg: &ExplosionMessage) {
         CTaskTurnGame::handle_message_raw(
             turn_game,
             sender,
-            TaskMessage::Explosion as u32,
+            TaskMessage::Explosion,
             core::mem::size_of::<ExplosionMessage>() as u32,
             msg as *const ExplosionMessage as *const u8,
         );
